third_party/WebKit/LayoutTests/editing/execCommand/clipboard-access-expected.txt

This test checks whether javascript programs can access clipboard.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS whenEnabled('Copy') is 0
PASS whenEnabled('Cut') is 0
PASS whenEnabled('Paste') is 0
PASS whenEnabled('PasteAndMatchStyle') is 0
PASS successfullyParsed is true

TEST COMPLETE
